refactor(hermes): seperate out domain & rpc models
